requirements.txt

django 
djangorestframework 
langgraph 
langchain 
langchain-google-genai 
python-dotenv
django-cors-headers